backend/requirements.txt
# FastAPI 웹 프레임워크 (비동기 처리 지원)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0  # ASGI 서버

# Flask 웹 프레임워크 (기존 버전 호환성 유지)
flask>=2.3.0
flask-cors>=4.0.0

# LangChain 관련 패키지 (LangChain 1.0)
langchain-core>=0.1.0
langchain-text-splitters>=0.0.1
langchain-community>=0.0.10

# 벡터 데이터베이스
chromadb>=0.4.0

# 임베딩 모델 및 머신러닝
sentence-transformers>=2.2.0

# 수치 계산 및 머신러닝 평가
numpy>=1.24.0
scikit-learn>=1.3.0

# Supabase PostgreSQL 연동
supabase>=2.0.0
python-dotenv>=1.0.0

# 유틸리티 (기본 내장 모듈이지만 명시)
# json, os, uuid, typing, datetime은 Python 표준 라이브러리

